# .bob/skills/agent-enablement/templates/openapi-tools.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Meridian AP Payments API",
    "version": "2.0.0",
    "description": "Modernized accounts-payable payments API. Read operations require an identity holding the 'inquiry' scope; write operations require the 'ops' scope. Scopes are derived from the caller's identity at the identity provider, not from anything in the request body."
  },
  "servers": [
    {
      "url": "https://<PUBLIC_URL>",
      "description": "Modernized AP payments service"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/v2/payments/{ref}": {
      "get": {
        "operationId": "payment_status_lookup",
        "summary": "Look up one AP payment by payment reference",
        "description": "Returns vendor, amount, status, hold reason, dates and risk for a single payment. Use when the caller has a payment reference such as MT-2026-08812.",
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "description": "Payment reference, e.g. MT-2026-08812",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "The payment",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Payment" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/IdentityUnverified" },
          "403": { "$ref": "#/components/responses/IdentityScopeDenied" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v2/payments": {
      "get": {
        "operationId": "payments_search",
        "summary": "Search AP payments",
        "description": "Search payments by invoice number, status, vendor or free text, with paging. Use when the caller does not have a payment reference in hand.",
        "parameters": [
          {
            "name": "invoice",
            "in": "query",
            "required": false,
            "description": "Invoice number, e.g. INV-2026-4471",
            "schema": { "type": "string" }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "Payment status filter",
            "schema": { "type": "string" }
          },
          {
            "name": "vendor",
            "in": "query",
            "required": false,
            "description": "Vendor name filter",
            "schema": { "type": "string" }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "description": "Free-text query across invoice, purchase order and vendor",
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page number, 1-based",
            "schema": { "type": "integer", "minimum": 1, "default": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of matching payments",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaymentPage" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/IdentityUnverified" },
          "403": { "$ref": "#/components/responses/IdentityScopeDenied" }
        }
      }
    },
    "/api/v2/payments/{ref}/release": {
      "post": {
        "operationId": "payment_release",
        "summary": "Release a held payment for settlement",
        "description": "Write operation. Requires an identity holding the payment-operations scope; an inquiry-only identity is refused with an auditable identity error.",
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "description": "Payment reference",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "note": {
                    "type": "string",
                    "description": "Reason recorded on the audit trail"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The payment after release",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Payment" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/IdentityUnverified" },
          "403": { "$ref": "#/components/responses/IdentityScopeDenied" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/v2/payments/{ref}/hold": {
      "post": {
        "operationId": "payment_hold",
        "summary": "Place a payment on hold",
        "description": "Write operation. Requires an identity holding the payment-operations scope; an inquiry-only identity is refused with an auditable identity error.",
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "description": "Payment reference",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "description": "Hold reason code or text"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The payment after being placed on hold",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Payment" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/IdentityUnverified" },
          "403": { "$ref": "#/components/responses/IdentityScopeDenied" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Identity token issued by the enterprise identity provider. The token's policies determine which operations are permitted; the API does not accept a role or scope supplied in the request."
      }
    },
    "responses": {
      "IdentityUnverified": {
        "description": "No verifiable caller identity",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/IdentityError" }
          }
        }
      },
      "IdentityScopeDenied": {
        "description": "A verified identity that does not hold the required scope",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/IdentityError" }
          }
        }
      },
      "NotFound": {
        "description": "No payment matches the supplied reference",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    },
    "schemas": {
      "Payment": {
        "type": "object",
        "properties": {
          "ref": { "type": "string", "description": "Payment reference" },
          "invoiceNo": { "type": "string" },
          "poNo": { "type": "string" },
          "vendor": {
            "type": "object",
            "properties": {
              "name": { "type": "string" },
              "country": { "type": "string" }
            }
          },
          "amount": {
            "type": "object",
            "properties": {
              "value": { "type": "number" },
              "currency": { "type": "string" }
            }
          },
          "status": { "type": "string" },
          "holdReason": { "type": "string", "nullable": true },
          "dates": {
            "type": "object",
            "properties": {
              "invoice": { "type": "string", "format": "date" },
              "due": { "type": "string", "format": "date" },
              "expectedPay": { "type": "string", "format": "date", "nullable": true }
            }
          },
          "risk": {
            "type": "object",
            "properties": {
              "score": { "type": "number" },
              "band": { "type": "string" }
            }
          }
        }
      },
      "PaymentPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Payment" }
          },
          "page": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "total": { "type": "integer" }
        }
      },
      "IdentityError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "identity_unverified or identity_scope_denied"
          },
          "detail": { "type": "string" },
          "policies": { "type": "array", "items": { "type": "string" } },
          "granted_scopes": { "type": "array", "items": { "type": "string" } },
          "required_scope": { "type": "string" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "detail": { "type": "string" }
        }
      }
    }
  }
}